requirements.txt
pynacl
keyring
SecretStorage
msgpack
requests
